ボタンの見た目変更、各画像の素材を追加
・追加：ボタン用の画像
・追加：タイマー用の画像
・追加：果物アイコン用の画像
・変更：CustomButtonView
・変更：ボタンのアニメーション
・削除：消し忘れていたデータ

+++ Team Game/Assets/Scripts/UI/Timer/Timer.cs
@@ -26,6 +26,6 @@ public class Timer : MonoBehaviour {
 		_time += Time.deltaTime;
 
 		Remaining = TimeLimit - (int)_time;
-		_slider.value = (float)TimeLimit - _time;
+		_slider.value = TimeLimit - _time;
 	}
 }
